refactored database to use abstraction
