+++ src/F3/Settings.php
@@ -3,13 +3,15 @@ namespace F3;
 
 class Settings {
 
-	const DB_HOST = '';
-	const DB_NAME = '';
-	const DB_USER = '';
-	const DB_PASS = '';
+	const DB_HOST = '@@DB_HOST@@';
+	const DB_NAME = '@@DB_NAME@@';
+	const DB_USER = '@@DB_USER@@';
+	const DB_PASS = '@@DB_PASS@@';
 	const DB_CHARSET = 'utf8';
 	
-	const GA_TRACKING_CODE = '';
+	const DB_LOCAL_PATH = 'db/f3sqlite.db';
+	
+	const GA_TRACKING_CODE = '@@GA_TRACKING_CODE@@';
 }
 
 ?>
